Store receipt sum as an int rather than a float

+++ backend/economicalc/objects/receipt.py
@@ -2,7 +2,7 @@ from datetime import date
 
 
 class Receipt:
-    def __init__(self, id: int, store: str, items: list, date_of_purchase: date, total_sum: float) -> None:
+    def __init__(self, id: int, store: str, items: list, date_of_purchase: date, total_sum: int) -> None:
         self.id = id
         self.store = store
         self.items = items
